Modified schema and restructured analysis task
