+++ core/src/db/entity/metadata/mod.rs
@@ -1,6 +1,7 @@
 mod common;
 mod media_metadata;
 pub(crate) mod prisma_macros;
+pub(crate) mod resolution;
 mod series_metadata;
 
 pub use common::{age_rating_deserializer, parse_age_restriction};
